B/fleet/requirements.txt
Flask
flask-sqlalchemy
mysql-connector
alembic
sqlalchemy-utils
pytz
werkzeug
unidecode
requests
gunicorn
newrelic